% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_censo.R
\name{get_censo}
\alias{get_censo}
\title{Get data from the 2021 Spanish census}
\usage{
get_censo(table, variables, metrics, language = "ES")
}
\arguments{
\item{table}{A string of length 1 with the table name to query data from.}

\item{variables}{A string with variable names of the queried table. If more than one, must be provided using c() (max. 4)}

\item{metrics}{A string with the units of measurement of the queried data. If more than one, must be provided concatenated using `c()`}

\item{language}{The language used in the body request (the only valid values are 'EN' or 'ES')}
}
\value{
A data.frame with the queried data
}
\description{
`get_censo` retrieves data from the 2021 Spanish census through the API of the National Institute of Statistics. For more info on the available tables and variables visit https://www.ine.es/dyngs/DataLab/en/manual.html?cid=1259945952385.
}
\examples{
df <- get_censo("hog", c("ID_SUP_VIV", "ID_RESIDENCIA_N1"), "SHOGARES")

}
