% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{metrics}
\alias{metrics}
\title{Available units of measurement}
\format{
A data frame with 51 rows and 4 columns:
\describe{
  \item{id_tabla}{Table ID}
  \item{id_unidad_de_medida}{unit ID}
  \item{descripcion_unidad_de_medida_es, descripcion_unidad_de_medida_en}{Content description of the unit of measurement in spanish and english}
}
}
\source{
<https://www.ine.es/censos2021/SDC21_variables_disponibles.xlsx>
}
\usage{
metrics
}
\description{
A list of the available units of measurement in the Spanish 2021 Census API
}
\keyword{datasets}
