% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-microdata.R
\name{read_microdata}
\alias{read_microdata}
\title{Read microdata from Censo 2010}
\usage{
read_microdata(
  microdata,
  topic = c("domicilios", "pessoas", "migracao", "mortalidade")
)
}
\arguments{
\item{microdata}{microdata file}

\item{topic}{which topic to read}
}
\value{
\code{data.frame()} with the results
}
\description{
Read microdata and parses using the awesome for operator
}
