% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tables}
\alias{tables}
\title{Available tables}
\format{
A data frame with 7 rows and 3 columns:
\describe{
  \item{id_tabla}{Table ID}
  \item{descripcion_tabla_es, descripcion_tabla_en}{Content description of the table in spanish and english}
}
}
\source{
<https://www.ine.es/censos2021/SDC21_tablas_disponibles.xlsx>
}
\usage{
tables
}
\description{
A list of the available tables in the Spanish 2021 Census API
}
\keyword{datasets}
