% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidycenso-package.R
\docType{package}
\name{tidycenso-package}
\alias{tidycenso}
\alias{tidycenso-package}
\title{tidycenso: Download Data From 2021 Spanish Census}
\description{
Retrieves data from the 2021 Spanish census through the National Institute of Statistics official API. For more info on the available tables and variables visit \url{https://www.ine.es/dyngs/DataLab/en/manual.html?cid=1259945952385}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ropenspain.github.io/tidycenso/}
  \item \url{https://github.com/rOpenSpain/tidycenso}
  \item Report bugs at \url{https://github.com/ropenspain/tidycenso/issues}
}

}
\keyword{internal}
