% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{variables}
\alias{variables}
\title{Available variables}
\format{
A data frame with 395 rows and 4 columns:
\describe{
  \item{id_tabla}{Table ID}
  \item{id_variable}{Variable ID}
  \item{descripcion_variable_es, descripcion_variable_en}{Content description of the variable in spanish and english}
}
}
\source{
<https://www.ine.es/censos2021/SDC21_variables_disponibles.xlsx>
}
\usage{
variables
}
\description{
A list of the available variables in the Spanish 2021 Census API
}
\keyword{datasets}
