% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_censo.R
\name{get_censo}
\alias{get_censo}
\title{Get data from the 2021 spanish census}
\usage{
get_censo(tabla, variables, metricas, idioma = "ES")
}
\arguments{
\item{tabla}{Name of the table to query data from.}

\item{variables}{The variable names of the queried table (max. 4)}

\item{metricas}{The units of measurement of the queried data.}

\item{idioma}{The language used in the request body (the only valid values are 'EN' or 'ES')}
}
\value{
A data.frame with the queried data
}
\description{
`get_censo` retrieves data from the 2021 spanish census through the API of the National Institute of Statistics. For more info on the available tables and variables visit https://www.ine.es/dyngs/DataLab/en/manual.html?cid=1259945952385.
}
\examples{
df <- get_censo("hog", c("ID_SUP_VIV", "ID_RESIDENCIA_N1"), "SHOGARES")

}
